pyproject.toml
[project]
name = "ai-insurance-claim-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "datasets>=3.0.0",
    "faiss-cpu>=1.8.0",
    "faker>=30.0.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.2.2",
    "ragas>=0.2.0",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.40.0",
    "torchvision>=0.26.0",
]
